requirements.txt
python-barcode
Pillow